% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_eevee_die_coords.R
\name{add_eevee_die_coords}
\alias{add_eevee_die_coords}
\title{Add EEVEE maskset die coordinates}
\usage{
add_eevee_die_coords(dat)
}
\arguments{
\item{dat}{A data frame containing EEVEE die identifications}
}
\value{
A data frame with containing the cell, bar, die ID and coordinates of every chip on the mask
}
\description{
Add EEVEE maskset die coordinates
}
