% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_eevee_grating_info.R
\name{add_eevee_grating_info}
\alias{add_eevee_grating_info}
\title{Add EEVEE grating design info to data frame}
\usage{
add_eevee_grating_info(dat)
}
\arguments{
\item{dat}{A data frame containing cell_id and bar_id columns}
}
\value{
A data frame with grt_no and grt_pitch columns added
}
\description{
Add EEVEE grating design info to data frame
}
