% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_eevee_ids.R
\name{generate_eevee_ids}
\alias{generate_eevee_ids}
\title{Generate all EEVEE cell-bar-die IDs}
\usage{
generate_eevee_ids()
}
\value{
A data frame of die identifications
}
\description{
Generate all EEVEE cell-bar-die IDs
}
\examples{
generate_eevee_ids()
}
