% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sivers_label_funcs.R
\name{generate_sivers_alphanumeric_sequence}
\alias{generate_sivers_alphanumeric_sequence}
\title{Generate Sivers' alpha-numeric chip labeling sequence}
\usage{
generate_sivers_alphanumeric_sequence()
}
\value{
A tibble with decimal and coded sequences
}
\description{
Generate Sivers' alpha-numeric chip labeling sequence
}
\examples{
sivers_labs = generate_sivers_alphanumeric_sequence()
}
