% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eevee_decoding_functions.R
\name{get_eevee_cell_id}
\alias{get_eevee_cell_id}
\title{Get EEVEE maskset cell ID}
\usage{
get_eevee_cell_id(bar_id, die_id)
}
\arguments{
\item{bar_id}{A two-element character vector}

\item{die_id}{A two-element character vector}
}
\value{
A two-element character vector with integers
}
\description{
Get EEVEE maskset cell ID
}
