% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eevee_decoding_functions.R
\name{get_eevee_die_coords}
\alias{get_eevee_die_coords}
\title{Get EEVEE maskset die coordinates}
\usage{
get_eevee_die_coords()
}
\value{
A data frame with containing the cell, bar, die ID and coordinates of every chip on the mask
}
\description{
Get EEVEE maskset die coordinates
}
