% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eevee_decoding_functions.R
\name{sivers_decimal_to_code}
\alias{sivers_decimal_to_code}
\title{Convert Sivers chip label from decimal to coded value}
\usage{
sivers_decimal_to_code(decimal)
}
\arguments{
\item{decimal}{A two-digit decimal integer}
}
\value{
A two-element character vector
}
\description{
Convert Sivers chip label from decimal to coded value
}
