% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sivers_label_funcs.R
\name{sivers_code_to_decimal}
\alias{sivers_code_to_decimal}
\title{Convert Sivers chip label from code to decimal value}
\usage{
sivers_code_to_decimal(code)
}
\arguments{
\item{code}{A two-element character vector}
}
\value{
A two-digit decimal integer
}
\description{
Convert Sivers chip label from code to decimal value
}
